Fixed several warnings

+++ src/algorithms/Entanglement.cpp
@@ -4,7 +4,7 @@ namespace qc {
     Entanglement::Entanglement(unsigned short nq) : QuantumComputation(nq) {
         emplace_back<StandardOperation>(nqubits, 0, H);
 
-        for(int i = 1; i < nq; i++) {
+        for(unsigned short i = 1; i < nq; i++) {
             emplace_back<StandardOperation>(nqubits, Control(0,Control::pos), i, X);
         }
     }
